cmd/internal/objabi: generalize "is runtime package" check
There are several implementations of "is this package path a runtime
package". They all have slightly different lists because they all care
about slightly different properties of building the runtime.

To start converging these, we replace objabi.IsRuntimePackagePath with
objabi.LookupPkgSpecial, which returns a struct we can extend with
various special build properties. We'll extend this with several other
flags in the following CLs.

Change-Id: I21959cb8c3d18a350d6060467681c72ea49af712