os: do not skip directory entries with zero inodes on wasip1
When building programs to GOOS=wasip1, the programs does not have the
guarantees that the underlying directories will come from typical
unix file systems where a zero inode value indicates that the entry
was deleted but not yet removed from the directory. The host runtime
may be running on windows or may be exposing virtual user-space file
systems that do not have the concept of inodes. In those setup, we
assume that the host runtime is in charge of dealing with edge cases
such as skipping directory entries with zero inodes when needed, and
the guest application should trust the list of entries that it sees;
therefore, we disable skipping over zero inodes on wasip1.

Change-Id: I99aa562441cdb4182965f270af054cf3cf7f8f20