debug/pe: add `Machine` type to align API with debug/elf and debug/macho
As noted in #60198, the API of `debug/pe` is slightly different to
`debug/macho` and `debug/elf`, due to not exposing a `Machine` type.

This means that it is slightly harder and less consistent across
the different packages, and we should ideally migrate to a consistent
interface.

This introduces the `Machine` type and associated helpers based on the
code that's found in `debug/macho`.

Closes #60198.

Change-Id: I432d1e664ccce7a3aa88a3ee09129d8e87300d1a