cmd/link: force ld64 when building non-PIE binary on macOS
Apple's new linker ld-prime silently ignores -no_pie and generates
a PIE binary. This doesn't work because in exe buildmode we assume
the binary is not relocated and so generate non-relocated
addresses, which will not be correct when the binary is actually
relocated.

(This is AMD64 Mac only, as it is always PIE on ARM64 Mac.)

For #61229.

Change-Id: I9d9f9f0395ba5db1e98209719a79eb0e395e1f4d