net/http: performance regression in http.ServeFile
The change to io.Copy in servefile caused a exponential increase in
execution time if the reader is a os.File.

I'm not sure what the issue is, if it is something related to io.Copy or
the network stack, but if the benchmark takes the io.Copy path instead of
CopyN it gets stuck and is killed after 10 minutes.

We could also revert the offending change: https://go-review.googlesource.com/c/go/+/446276

```
BenchmarkServeFile-16              25746             46767 ns/op
```

Fixes: https://github.com/golang/go/issues/61530

Change-Id: I807259cf5c9a7b728ebaeeb219c868d5f5567119
GitHub-Last-Rev: 0a07d04ccaf82e706b2550d7b91734ed2451a1b9
GitHub-Pull-Request: golang/go#61532