cmd/compile: try to rewrite loops to count down
This reduce the pressure on regalloc because then the loop only keep alive
one value (the iterator) instead of the iterator and the upper bound since
the comparison now acts against an immediate, often zero which can be skipped.

This optimize things like the the proposed "for range N" syntax from #61405.

Change-Id: I5facff8b267a0b60ea3c1b9a58c4d74cdb38f03f