archive/tar: add AddFS method to Writer
The method AddFS can be used to add the contents of a fs.FS filesystem to a tar archive. This method walks the directory tree starting at the root of the filesystem and adds each file to the archive.

Fixes #58000

Change-Id: I0e6abac76b76bc5d95b74b7e5b37634ce3f76c85
GitHub-Last-Rev: 95733dcc8bef41a8f35819914935b4ea01aab387
GitHub-Pull-Request: golang/go#61599