Update patch set 3
Abandoned

I didn't end up switching to --map-current-user, so this isn't needed for that reason anymore. Instead, CL 513779 was used to make other tests compatible with our use of --map-root-user.

It's still possible these tests would be fine to skip in -short test mode, but that is best left for a more intentional and holistic CL.

Patch-set: 3
Status: abandoned
Tag: autogenerated:gerrit:abandon