all: add GOARMFP env var for ARM floating point mode
This change introduces a new environment variable, GOARMFP,
which controls the floating point mode on ARM targets. The
GOARMFP flag can be set to 'hard' (default) or 'soft', to
select whether to use hardware instructions or software
emulation for floating point computations, respectively.

Previously, software floating point support was limited to
GOARM=5. With the introduction of GOARMFP, software floating
point is now extended to all ARM versions, including GOARM=6
and 7. This avoids the need to downgrade to the ARMv5
instruction set on ARMv7 chips without hardware FPUs.

When GOARM=5, only soft floating point is used, regardless of
the GOARMFP value.

For #61588

Change-Id: I23dc86fbd0733b262004a2ed001e1032cf371e94