runtime: avoid stack unwinding in deferreturn
A consequence of go.dev/cl/513837 was that we started using the
unwinder to find the current stack frame in deferreturn. It turns out
this is actually a significant performance impact for some workloads.

This CL changes deferreturn to never look up the current stack frame.
This isn't a problem for functions that use linked defers, because
they only care about comparing p.sp to d.sp.

But if there are pending open-coded defers in the frame, we need to
explicitly push a pseudo defer entry to carry that extra state so
deferreturn (in _panic.start) can restore it.

Change-Id: I11fa73649302199eadccc27b403b231db8f33db2