runtime: add deferreturn fast path for linked defers
A consequence of go.dev/cl/513837 was that functions using linked
defers would now always need to unwind the current frame in
deferreturn, even though they have no use for the extra data retrieved
this way. Further, it turns out there are workloads where this has a
significant performance impact.

This CL adds a fast path for deferreturn to better handle these cases,
by short-circuiting the extra lookups.

Notably, this doesn't handle the corner case of calling deferreturn in
a function that uses linked defer when dynamically there just aren't
any defers pending. However, that doesn't seem to be a problem for the
affected workloads, and I hope to address it in the future with
further defer optimizations.

Change-Id: I11fa73649302199eadccc27b403b231db8f33db2