cmd/internal/obj/arm64: allows reporting multiple errors while encoding
The arm64 new assembler is based on heuristic matching of machine
instructions. When a Prog is successfully matched by matching multiple
machine instructions, the previous unsuccessful match is likely to
report an error message, but because the final match is successful, so
the previous error messages are bogus and we need to delete them. This
CL hijacks ctxt7.ctxt.DiagFunc into a local function mayDiag before
encoding. This function caches errors reported during encoding in a
buffer. If the final match is successful, the cache will be cleared,
otherwise the cached error will be flushed.

At the same time, since each element of the instruction is encoded
separately, we may have added necessary error checking code when
encoding each element, so multiple error messages may be reported when
encoding an instruction is unsuccessful. We have no control over
reporting only one error because, depending on the cases, we cannot
tell which error should be reported and which error should not be
reported. So this CL also makes some changes to testErrors to allow
multiple errors for one file line, and the test is considered
successful as long as the expected error is found.

Updates #44734

Change-Id: Ib8e47f87ed564b42304eb6ee7ec08687ae05756c