cmd/asm: allows shrinking of error messages reported in the assembler
The arm64 new assembler is based on heuristic matching of machine
instructions. When a Prog is successfully matched by matching multiple
machine instructions, the previous unsuccessful match is likely to
report an error message, but because the final match is successful, so
the previous error messages are bogus and we need to delete them. This
CL adds a DiagShrink function field to the obj.Link structure to allow
removing logged error messages. This CL has the following changes:

For gc, add a DiagShrink function.

For asm and testEndToEnd:
1. In DiagFunc, save error messages in a slice instead of reporting
errors immediately.
2. Add a DiagFlush function to report errors.
3. Add a DiagShrink function.

For testErrors:
1. In DiagFunc, save the errors captured during parsing in errBuf, and
save the other errors in a slice.
2. Add a DiagFlush function to write the error message in the slice into
errBuf.
3. Add a DiagShrink function.
4. Multiple errors are allowed for a file line, and the test is
considered successful as long as the expected error matches.

Updates #44734

Change-Id: Ib8e47f87ed564b42304eb6ee7ec08687ae05756c