archive/tar: add AddFS method to Writer
The method AddFS can be used to add the contents of a fs.FS filesystem to a tar archive. This method walks the directory tree starting at the root of the filesystem and adds each file to the archive.

DO NOT REVIEW
DO NOT SUBMIT

Fixes #58000

Change-Id: I0e6abac76b76bc5d95b74b7e5b37634ce3f76c85
GitHub-Last-Rev: ac0535790170cbf1ad85fb6809c99d99da263a17
GitHub-Pull-Request: golang/go#61599