cmd/dist: omit -trimpath if GOROOT_FINAL is set
DO NOT MERGE

With this change, any build environment that sets GOROOT_FINAL
(which was, up until Go 1.21, probably most OS distributions of the GO
toolchain) would end up with binaries that depend on their choice of
GOROOT_FINAL, making it less likely that their binaries will match the
(otherwise reproducible) build distributed by the Go project.

Moreover, we could get rid of a fair amount of complexity in the Go
codebase if we could eliminate GOROOT_FINAL entirely instead of
restoring it.

Let's hold off on this until we see whether it is really necessary.

For #61921.

Change-Id: Ib93d10b59ca6326472e28d6234282d4472cfa529