types2: implement type-checking of predeclared zero
This change introduces a new (unexported for now) basic type,
_UntypedZero, which is used to represent the type of the pre-
declared zero value.

The CL also introduces a new Object kind, _Value (also currently
unexported), which represents the predeclared values nil and
zero. For backward-compatibility, _Value is an alias for Nil.

Internally, a zero operand uses the new zerovalue operand mode,
to represent a possibly typed zero value, e.g., after a zero
value is implicitly converted to an array or a struct type.
By having a special operand mode, we keep track of the fact
that the value is zero, even if the type has changed.
This is matching the code for untyped nil in types2.

(Note that untyped nils are handled slightly differently in
go/types for backward-compatibility with external tools such
as go vet.)

The remaining changes closely follow the respespective code
for untyped nils.

Finally, a new error code, UntypedZeroUse, is introduced and
used for certain misuses of zero.

For #61372.

Change-Id: I46824147ea5371e107cef4731fa87e89eb7d01a7